LayoutTests/imported/w3c/web-platform-tests/mathml/relations/css-styling/mathvariant-auto-selection-expected.txt


FAIL Selection.toString() for <mi>A</mi> returns the transformed character. assert_equals: expected "ğ´" but got "A"
FAIL Selection.toString() for <mi>B</mi> returns the transformed character. assert_equals: expected "ğµ" but got "B"
FAIL Selection.toString() for <mi>C</mi> returns the transformed character. assert_equals: expected "ğ¶" but got "C"
FAIL Selection.toString() for <mi>D</mi> returns the transformed character. assert_equals: expected "ğ·" but got "D"
FAIL Selection.toString() for <mi>E</mi> returns the transformed character. assert_equals: expected "ğ¸" but got "E"
FAIL Selection.toString() for <mi>F</mi> returns the transformed character. assert_equals: expected "ğ¹" but got "F"
FAIL Selection.toString() for <mi>G</mi> returns the transformed character. assert_equals: expected "ğº" but got "G"
FAIL Selection.toString() for <mi>H</mi> returns the transformed character. assert_equals: expected "ğ»" but got "H"
FAIL Selection.toString() for <mi>I</mi> returns the transformed character. assert_equals: expected "ğ¼" but got "I"
FAIL Selection.toString() for <mi>J</mi> returns the transformed character. assert_equals: expected "ğ½" but got "J"
FAIL Selection.toString() for <mi>K</mi> returns the transformed character. assert_equals: expected "ğ¾" but got "K"
FAIL Selection.toString() for <mi>L</mi> returns the transformed character. assert_equals: expected "ğ¿" but got "L"
FAIL Selection.toString() for <mi>M</mi> returns the transformed character. assert_equals: expected "ğ‘€" but got "M"
FAIL Selection.toString() for <mi>N</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "N"
FAIL Selection.toString() for <mi>O</mi> returns the transformed character. assert_equals: expected "ğ‘‚" but got "O"
FAIL Selection.toString() for <mi>P</mi> returns the transformed character. assert_equals: expected "ğ‘ƒ" but got "P"
FAIL Selection.toString() for <mi>Q</mi> returns the transformed character. assert_equals: expected "ğ‘„" but got "Q"
FAIL Selection.toString() for <mi>R</mi> returns the transformed character. assert_equals: expected "ğ‘…" but got "R"
FAIL Selection.toString() for <mi>S</mi> returns the transformed character. assert_equals: expected "ğ‘†" but got "S"
FAIL Selection.toString() for <mi>T</mi> returns the transformed character. assert_equals: expected "ğ‘‡" but got "T"
FAIL Selection.toString() for <mi>U</mi> returns the transformed character. assert_equals: expected "ğ‘ˆ" but got "U"
FAIL Selection.toString() for <mi>V</mi> returns the transformed character. assert_equals: expected "ğ‘‰" but got "V"
FAIL Selection.toString() for <mi>W</mi> returns the transformed character. assert_equals: expected "ğ‘Š" but got "W"
FAIL Selection.toString() for <mi>X</mi> returns the transformed character. assert_equals: expected "ğ‘‹" but got "X"
FAIL Selection.toString() for <mi>Y</mi> returns the transformed character. assert_equals: expected "ğ‘Œ" but got "Y"
FAIL Selection.toString() for <mi>Z</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "Z"
FAIL Selection.toString() for <mi>a</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "a"
FAIL Selection.toString() for <mi>b</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "b"
FAIL Selection.toString() for <mi>c</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "c"
FAIL Selection.toString() for <mi>d</mi> returns the transformed character. assert_equals: expected "ğ‘‘" but got "d"
FAIL Selection.toString() for <mi>e</mi> returns the transformed character. assert_equals: expected "ğ‘’" but got "e"
FAIL Selection.toString() for <mi>f</mi> returns the transformed character. assert_equals: expected "ğ‘“" but got "f"
FAIL Selection.toString() for <mi>g</mi> returns the transformed character. assert_equals: expected "ğ‘”" but got "g"
FAIL Selection.toString() for <mi>h</mi> returns the transformed character. assert_equals: expected "â„" but got "h"
FAIL Selection.toString() for <mi>i</mi> returns the transformed character. assert_equals: expected "ğ‘–" but got "i"
FAIL Selection.toString() for <mi>j</mi> returns the transformed character. assert_equals: expected "ğ‘—" but got "j"
FAIL Selection.toString() for <mi>k</mi> returns the transformed character. assert_equals: expected "ğ‘˜" but got "k"
FAIL Selection.toString() for <mi>l</mi> returns the transformed character. assert_equals: expected "ğ‘™" but got "l"
FAIL Selection.toString() for <mi>m</mi> returns the transformed character. assert_equals: expected "ğ‘š" but got "m"
FAIL Selection.toString() for <mi>n</mi> returns the transformed character. assert_equals: expected "ğ‘›" but got "n"
FAIL Selection.toString() for <mi>o</mi> returns the transformed character. assert_equals: expected "ğ‘œ" but got "o"
FAIL Selection.toString() for <mi>p</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "p"
FAIL Selection.toString() for <mi>q</mi> returns the transformed character. assert_equals: expected "ğ‘" but got "q"
FAIL Selection.toString() for <mi>r</mi> returns the transformed character. assert_equals: expected "ğ‘Ÿ" but got "r"
FAIL Selection.toString() for <mi>s</mi> returns the transformed character. assert_equals: expected "ğ‘ " but got "s"
FAIL Selection.toString() for <mi>t</mi> returns the transformed character. assert_equals: expected "ğ‘¡" but got "t"
FAIL Selection.toString() for <mi>u</mi> returns the transformed character. assert_equals: expected "ğ‘¢" but got "u"
FAIL Selection.toString() for <mi>v</mi> returns the transformed character. assert_equals: expected "ğ‘£" but got "v"
FAIL Selection.toString() for <mi>w</mi> returns the transformed character. assert_equals: expected "ğ‘¤" but got "w"
FAIL Selection.toString() for <mi>x</mi> returns the transformed character. assert_equals: expected "ğ‘¥" but got "x"
FAIL Selection.toString() for <mi>y</mi> returns the transformed character. assert_equals: expected "ğ‘¦" but got "y"
FAIL Selection.toString() for <mi>z</mi> returns the transformed character. assert_equals: expected "ğ‘§" but got "z"
FAIL Selection.toString() for <mi>Ä±</mi> returns the transformed character. assert_equals: expected "ğš¤" but got "Ä±"
FAIL Selection.toString() for <mi>È·</mi> returns the transformed character. assert_equals: expected "ğš¥" but got "È·"
FAIL Selection.toString() for <mi>Î‘</mi> returns the transformed character. assert_equals: expected "ğ›¢" but got "Î‘"
FAIL Selection.toString() for <mi>Î’</mi> returns the transformed character. assert_equals: expected "ğ›£" but got "Î’"
FAIL Selection.toString() for <mi>Î“</mi> returns the transformed character. assert_equals: expected "ğ›¤" but got "Î“"
FAIL Selection.toString() for <mi>Î”</mi> returns the transformed character. assert_equals: expected "ğ›¥" but got "Î”"
FAIL Selection.toString() for <mi>Î•</mi> returns the transformed character. assert_equals: expected "ğ›¦" but got "Î•"
FAIL Selection.toString() for <mi>Î–</mi> returns the transformed character. assert_equals: expected "ğ›§" but got "Î–"
FAIL Selection.toString() for <mi>Î—</mi> returns the transformed character. assert_equals: expected "ğ›¨" but got "Î—"
FAIL Selection.toString() for <mi>Î˜</mi> returns the transformed character. assert_equals: expected "ğ›©" but got "Î˜"
FAIL Selection.toString() for <mi>Î™</mi> returns the transformed character. assert_equals: expected "ğ›ª" but got "Î™"
FAIL Selection.toString() for <mi>Îš</mi> returns the transformed character. assert_equals: expected "ğ›«" but got "Îš"
FAIL Selection.toString() for <mi>Î›</mi> returns the transformed character. assert_equals: expected "ğ›¬" but got "Î›"
FAIL Selection.toString() for <mi>Îœ</mi> returns the transformed character. assert_equals: expected "ğ›­" but got "Îœ"
FAIL Selection.toString() for <mi>Î</mi> returns the transformed character. assert_equals: expected "ğ›®" but got "Î"
FAIL Selection.toString() for <mi>Î</mi> returns the transformed character. assert_equals: expected "ğ›¯" but got "Î"
FAIL Selection.toString() for <mi>ÎŸ</mi> returns the transformed character. assert_equals: expected "ğ›°" but got "ÎŸ"
FAIL Selection.toString() for <mi>Î </mi> returns the transformed character. assert_equals: expected "ğ›±" but got "Î "
FAIL Selection.toString() for <mi>Î¡</mi> returns the transformed character. assert_equals: expected "ğ›²" but got "Î¡"
FAIL Selection.toString() for <mi>Î£</mi> returns the transformed character. assert_equals: expected "ğ›´" but got "Î£"
FAIL Selection.toString() for <mi>Î¤</mi> returns the transformed character. assert_equals: expected "ğ›µ" but got "Î¤"
FAIL Selection.toString() for <mi>Î¥</mi> returns the transformed character. assert_equals: expected "ğ›¶" but got "Î¥"
FAIL Selection.toString() for <mi>Î¦</mi> returns the transformed character. assert_equals: expected "ğ›·" but got "Î¦"
FAIL Selection.toString() for <mi>Î§</mi> returns the transformed character. assert_equals: expected "ğ›¸" but got "Î§"
FAIL Selection.toString() for <mi>Î¨</mi> returns the transformed character. assert_equals: expected "ğ›¹" but got "Î¨"
FAIL Selection.toString() for <mi>Î©</mi> returns the transformed character. assert_equals: expected "ğ›º" but got "Î©"
FAIL Selection.toString() for <mi>Î±</mi> returns the transformed character. assert_equals: expected "ğ›¼" but got "Î±"
FAIL Selection.toString() for <mi>Î²</mi> returns the transformed character. assert_equals: expected "ğ›½" but got "Î²"
FAIL Selection.toString() for <mi>Î³</mi> returns the transformed character. assert_equals: expected "ğ›¾" but got "Î³"
FAIL Selection.toString() for <mi>Î´</mi> returns the transformed character. assert_equals: expected "ğ›¿" but got "Î´"
FAIL Selection.toString() for <mi>Îµ</mi> returns the transformed character. assert_equals: expected "ğœ€" but got "Îµ"
FAIL Selection.toString() for <mi>Î¶</mi> returns the transformed character. assert_equals: expected "ğœ" but got "Î¶"
FAIL Selection.toString() for <mi>Î·</mi> returns the transformed character. assert_equals: expected "ğœ‚" but got "Î·"
FAIL Selection.toString() for <mi>Î¸</mi> returns the transformed character. assert_equals: expected "ğœƒ" but got "Î¸"
FAIL Selection.toString() for <mi>Î¹</mi> returns the transformed character. assert_equals: expected "ğœ„" but got "Î¹"
FAIL Selection.toString() for <mi>Îº</mi> returns the transformed character. assert_equals: expected "ğœ…" but got "Îº"
FAIL Selection.toString() for <mi>Î»</mi> returns the transformed character. assert_equals: expected "ğœ†" but got "Î»"
FAIL Selection.toString() for <mi>Î¼</mi> returns the transformed character. assert_equals: expected "ğœ‡" but got "Î¼"
FAIL Selection.toString() for <mi>Î½</mi> returns the transformed character. assert_equals: expected "ğœˆ" but got "Î½"
FAIL Selection.toString() for <mi>Î¾</mi> returns the transformed character. assert_equals: expected "ğœ‰" but got "Î¾"
FAIL Selection.toString() for <mi>Î¿</mi> returns the transformed character. assert_equals: expected "ğœŠ" but got "Î¿"
FAIL Selection.toString() for <mi>Ï€</mi> returns the transformed character. assert_equals: expected "ğœ‹" but got "Ï€"
FAIL Selection.toString() for <mi>Ï</mi> returns the transformed character. assert_equals: expected "ğœŒ" but got "Ï"
FAIL Selection.toString() for <mi>Ï‚</mi> returns the transformed character. assert_equals: expected "ğœ" but got "Ï‚"
FAIL Selection.toString() for <mi>Ïƒ</mi> returns the transformed character. assert_equals: expected "ğœ" but got "Ïƒ"
FAIL Selection.toString() for <mi>Ï„</mi> returns the transformed character. assert_equals: expected "ğœ" but got "Ï„"
FAIL Selection.toString() for <mi>Ï…</mi> returns the transformed character. assert_equals: expected "ğœ" but got "Ï…"
FAIL Selection.toString() for <mi>Ï†</mi> returns the transformed character. assert_equals: expected "ğœ‘" but got "Ï†"
FAIL Selection.toString() for <mi>Ï‡</mi> returns the transformed character. assert_equals: expected "ğœ’" but got "Ï‡"
FAIL Selection.toString() for <mi>Ïˆ</mi> returns the transformed character. assert_equals: expected "ğœ“" but got "Ïˆ"
FAIL Selection.toString() for <mi>Ï‰</mi> returns the transformed character. assert_equals: expected "ğœ”" but got "Ï‰"
FAIL Selection.toString() for <mi>Ï‘</mi> returns the transformed character. assert_equals: expected "ğœ—" but got "Ï‘"
FAIL Selection.toString() for <mi>Ï•</mi> returns the transformed character. assert_equals: expected "ğœ™" but got "Ï•"
FAIL Selection.toString() for <mi>Ï–</mi> returns the transformed character. assert_equals: expected "ğœ›" but got "Ï–"
FAIL Selection.toString() for <mi>Ï°</mi> returns the transformed character. assert_equals: expected "ğœ˜" but got "Ï°"
FAIL Selection.toString() for <mi>Ï±</mi> returns the transformed character. assert_equals: expected "ğœš" but got "Ï±"
FAIL Selection.toString() for <mi>Ï´</mi> returns the transformed character. assert_equals: expected "ğ›³" but got "Ï´"
FAIL Selection.toString() for <mi>Ïµ</mi> returns the transformed character. assert_equals: expected "ğœ–" but got "Ïµ"
FAIL Selection.toString() for <mi>âˆ‚</mi> returns the transformed character. assert_equals: expected "ğœ•" but got "âˆ‚"
FAIL Selection.toString() for <mi>âˆ‡</mi> returns the transformed character. assert_equals: expected "ğ›»" but got "âˆ‡"
